add bio.tools to the main function and outline the structure with missing TODOs
